Started 'Rock Paper Scissors' Script
I started making the 'Rock, Paper, Scissors' script by linking an html
file to a JavaScript file in order to run a script for the game. I
created a function that take in 2 parameters: The player's selection
(from a prompt) and the computer's selection (floored random number
times the length of the list ['Rock', 'Paper', 'Scissors]).